requirements.txt
Django==1.9.6
uWSGI>=2.0.18,<2.1
django-admin-bootstrapped-compatible
django-extensions==1.6.7
django-nose==1.4.3
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
psycopg2==2.8.4
sorl-thumbnail==12.4.1
raven
dj-database-url
dj_database_url
django-storages==1.6
django-admin-rangefilter
arrow
sendinblue
babel
pillow==6.2.2
boto3
